+++ Cargo.toml
@@ -14,6 +14,7 @@ members = [
     "chain-signatures/chain-solana",
     "chain-signatures/chain-ethereum",
     "chain-signatures/chain-near",
+    "chain-signatures/utils",
 ]
 exclude = ["infra/scripts/generate_keys"]
 resolver = "2"
@@ -94,6 +95,7 @@ mpc-chain-ethereum = { path = "./chain-signatures/chain-ethereum" }
 mpc-chain-solana = { path = "./chain-signatures/chain-solana" }
 mpc-chain-near = { path = "./chain-signatures/chain-near" }
 mpc-keys = { path = "./chain-signatures/keys" }
+mpc-utils = { path = "./chain-signatures/utils" }
 mpc-node = { path = "./chain-signatures/node" }
 mpc-primitives = { path = "./chain-signatures/primitives" }
 signet-crypto = { path = "./signet-crypto" }
