+++ chain-signatures/chain-ethereum/Cargo.toml
@@ -21,6 +21,7 @@ k256.workspace = true
 mpc-chain-integration-core.workspace = true
 mpc-crypto.workspace = true
 mpc-primitives.workspace = true
+mpc-utils.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tokio.workspace = true
