+++ chain-signatures/chain-ethereum/src/finalized_head.rs
@@ -4,8 +4,8 @@ use crate::config::IndexerConfig;
 use crate::EthConfig;
 use alloy::eips::BlockNumberOrTag;
 use alloy::rpc::types::BlockId;
-use mpc_chain_integration_core::utils::task::{AbortOnDrop, CancellationTokenExt};
 use mpc_primitives::{Chain, ChainConfig as _};
+use mpc_utils::task::{AbortOnDrop, CancellationTokenExt};
 use std::sync::Arc;
 use std::time::{Duration, Instant};
 use tokio::sync::watch;
