+++ chain-signatures/chain-ethereum/src/indexer.rs
@@ -10,9 +10,9 @@ use alloy::sol_types::SolEvent;
 use anyhow::Context as _;
 use async_trait::async_trait;
 use futures_util::{stream, Stream};
-use mpc_chain_integration_core::utils::task::{retry_until_ok, retry_until_some, AbortOnDrop};
 use mpc_chain_integration_core::{ChainIndexer, ChainTelemetry, StateManager};
 use mpc_primitives::{Chain, ChainEvent, IndexedSignRequest};
+use mpc_utils::task::{retry_until_ok, retry_until_some, AbortOnDrop};
 use std::sync::{Arc, Mutex};
 use tokio::sync::mpsc;
 use tokio::time::Duration;
