+++ chain-signatures/chain-ethereum/src/lib.rs
@@ -16,7 +16,6 @@ pub mod publisher;
 mod respond_bidirectional;
 #[cfg(test)]
 mod test_utils;
-mod util;
 
 pub use client::{CatchupItem, MaybeBlock};
 pub use config::{EthConfig, GasConfig, IndexerConfig, PublisherConfig, RpcConfig};
