+++ chain-signatures/chain-near/src/indexer.rs
@@ -1,13 +1,12 @@
 use mpc_chain_integration_core::StateManager;
 use mpc_contract::primitives::PendingRequest;
 use mpc_primitives::{Chain, IndexedSignRequest, SignArgs, SignCommand, SignId};
+use mpc_utils::current_unix_timestamp;
 use near_account_id::AccountId;
 use std::collections::{HashMap, HashSet};
 use std::time::{Duration, Instant};
 use tokio::sync::mpsc;
 use tokio::task::JoinHandle;
-
-use crate::util::current_unix_timestamp;
 
 /// Configures the NEAR indexer.
 #[derive(Debug, Clone, clap::Parser)]
