+++ chain-signatures/chain-near/src/util.rs
@@ -17,15 +17,6 @@ impl AffinePointExt for AffinePoint {
     }
 }
 
-/// Current time in seconds since the UNIX epoch.
-pub fn current_unix_timestamp() -> u64 {
-    use std::time::{SystemTime, UNIX_EPOCH};
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("Time went backwards")
-        .as_secs()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
