+++ chain-signatures/chain-solana/Cargo.toml
@@ -18,6 +18,7 @@ k256.workspace = true
 mpc-chain-integration-core.workspace = true
 mpc-crypto.workspace = true
 mpc-primitives.workspace = true
+mpc-utils.workspace = true
 reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
