+++ chain-signatures/chain-solana/src/utils.rs
@@ -1,5 +1,3 @@
-use std::time::{SystemTime, UNIX_EPOCH};
-
 pub fn mpc_to_sol_signature(
     signature: &mpc_primitives::Signature,
     big_r: k256::EncodedPoint,
@@ -12,12 +10,4 @@ pub fn mpc_to_sol_signature(
         s: signature.s.to_bytes().into(),
         recovery_id: signature.recovery_id,
     }
-}
-
-// TODO: this is a duplicate of the same function in `mpc-node`, consider moving to shared crate like `mpc-utils` or something.
-pub fn current_unix_timestamp() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("Time went backwards")
-        .as_secs()
 }
