+++ chain-signatures/node/Cargo.toml
@@ -87,6 +87,7 @@ mpc-chain-solana.workspace = true
 mpc-chain-near.workspace = true
 mpc-keys.workspace = true
 mpc-primitives.workspace = true
+mpc-utils.workspace = true
 
 alloy-dyn-abi.workspace = true
 alloy-json-abi.workspace = true
