+++ chain-signatures/node/src/indexer_hydration/mod.rs
@@ -23,6 +23,7 @@ use mpc_primitives::{
     SignCommand, SignId, Signature, SignatureRespondedEvent, LATEST_MPC_KEY_VERSION,
     MAX_SECP256K1_SCALAR,
 };
+use mpc_utils::current_unix_timestamp;
 use sp_core::crypto::{AccountId32 as SpAccountId32, Ss58AddressFormatRegistry, Ss58Codec};
 use sp_core::{twox_128, H256};
 use sp_runtime::traits::BlakeTwo256;
@@ -109,7 +110,7 @@ impl HydrationSignatureRequestedEvent {
                 key_version: self.key_version,
             },
             Chain::Hydration,
-            crate::util::current_unix_timestamp(),
+            current_unix_timestamp(),
         ))
     }
 
@@ -235,7 +236,7 @@ impl HydrationSignBidirectionalRequestedEvent {
                 key_version: self.key_version,
             },
             Chain::Hydration,
-            crate::util::current_unix_timestamp(),
+            current_unix_timestamp(),
             SignBidirectionalEvent {
                 sender: self.sender,
                 serialized_transaction: self.serialized_transaction.clone(),
