+++ chain-signatures/utils/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "mpc-utils"
+version.workspace = true
+edition = "2021"
+description = "Generic shared utilities (time, task, retry) for the MPC workspace."
+
+[dependencies]
+anyhow.workspace = true
+backon.workspace = true
+rand.workspace = true
+tokio.workspace = true
+tokio-util.workspace = true
+tracing.workspace = true
