+++ chain-signatures/utils/src/lib.rs
@@ -1,0 +1,8 @@
+//! Generic shared utilities for the MPC workspace (time, task, retry).
+//! Domain-specific helpers live in `mpc-chain-integration-core`.
+
+pub mod retry;
+pub mod task;
+pub mod time;
+
+pub use time::current_unix_timestamp;
