+++ chain-signatures/utils/src/retry.rs
@@ -1,0 +1,262 @@
+use std::sync::atomic::{AtomicU32, AtomicU64, Ordering};
+use std::sync::Arc;
+use std::time::{Duration, SystemTime, UNIX_EPOCH};
+
+use backon::ExponentialBuilder;
+use rand::Rng;
+
+const RATE_LIMIT_BASE_COOLDOWN: Duration = Duration::from_secs(1);
+const RATE_LIMIT_MAX_COOLDOWN: Duration = Duration::from_secs(60);
+const RATE_LIMIT_MAX_JITTER: Duration = Duration::from_millis(500);
+const MAX_PENALTY_LEVEL: u32 = 10;
+
+/// Shared, 429-aware cooldown gate.
+#[derive(Clone)]
+pub struct SharedBackoff {
+    /// Inner state is shared across all clones of this instance to avoid per-instance fragmentation
+    inner: Arc<SharedBackoffInner>,
+}
+
+struct SharedBackoffInner {
+    /// Epoch millis until which all requests should wait.
+    limited_until_ms: AtomicU64,
+    /// Exponential penalty level, used to compute the cooldown duration.
+    penalty_level: AtomicU32,
+    /// Base cooldown duration for the first penalty level.
+    base_cooldown: Duration,
+    /// Maximum cooldown duration for the highest penalty level.
+    max_cooldown: Duration,
+}
+
+impl Default for SharedBackoff {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl SharedBackoff {
+    pub fn new() -> Self {
+        Self::with_cooldowns(RATE_LIMIT_BASE_COOLDOWN, RATE_LIMIT_MAX_COOLDOWN)
+    }
+
+    pub fn with_cooldowns(base_cooldown: Duration, max_cooldown: Duration) -> Self {
+        Self {
+            inner: Arc::new(SharedBackoffInner {
+                limited_until_ms: AtomicU64::new(0),
+                penalty_level: AtomicU32::new(0),
+                base_cooldown,
+                max_cooldown,
+            }),
+        }
+    }
+
+    /// Extends the global cooldown window after an observed 429. Concurrent
+    /// reports grow the penalty but only ever extend the window, never shorten
+    /// it. Returns the cooldown that was applied.
+    pub fn report_rate_limited(&self) -> Duration {
+        let level = self.inner.penalty_level.fetch_add(1, Ordering::Relaxed);
+        self.inner
+            .penalty_level
+            .fetch_min(MAX_PENALTY_LEVEL, Ordering::Relaxed);
+        let cooldown = self.cooldown_for_level(level);
+        self.inner
+            .limited_until_ms
+            .fetch_max(now_ms() + cooldown.as_millis() as u64, Ordering::Relaxed);
+        cooldown
+    }
+
+    /// Resets the penalty level after a successful call.
+    pub fn report_success(&self) {
+        self.inner.penalty_level.store(0, Ordering::Relaxed);
+    }
+
+    /// Sleeps until the global cooldown window has elapsed, plus random jitter
+    /// so callers parked on the same window don't all fire at once.
+    pub async fn wait(&self) {
+        let remaining = self.remaining();
+        if remaining.is_zero() {
+            return;
+        }
+        let jitter_cap = (remaining / 2).min(RATE_LIMIT_MAX_JITTER);
+        let jitter = rand::thread_rng().gen_range(Duration::ZERO..=jitter_cap);
+        tokio::time::sleep(remaining + jitter).await;
+    }
+
+    /// Time left in the current cooldown window, zero if not limited.
+    pub fn remaining(&self) -> Duration {
+        let ms = self
+            .inner
+            .limited_until_ms
+            .load(Ordering::Relaxed)
+            .saturating_sub(now_ms());
+        Duration::from_millis(ms)
+    }
+
+    fn cooldown_for_level(&self, level: u32) -> Duration {
+        let millis = (self.inner.base_cooldown.as_millis() as u64)
+            .saturating_mul(1u64 << level.min(MAX_PENALTY_LEVEL));
+        Duration::from_millis(millis).min(self.inner.max_cooldown)
+    }
+}
+
+fn now_ms() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_millis() as u64
+}
+
+/// Returns true if the error looks like an HTTP 429 (rate limited).
+pub fn is_rate_limited(e: &anyhow::Error) -> bool {
+    contains_status_code(&e.to_string(), "429")
+}
+
+/// Configuration for retrying RPC calls with exponential backoff.
+#[derive(Clone, Copy, Debug)]
+pub struct RetryConfig {
+    pub min_delay: Duration,
+    pub max_delay: Duration,
+    pub max_times: usize,
+    pub jitter: bool,
+}
+
+impl RetryConfig {
+    /// Builds an [`ExponentialBuilder`] from this configuration.
+    pub fn build(self) -> ExponentialBuilder {
+        let mut b = ExponentialBuilder::default()
+            .with_min_delay(self.min_delay)
+            .with_max_delay(self.max_delay)
+            .with_max_times(self.max_times);
+        if self.jitter {
+            b = b.with_jitter();
+        }
+        b
+    }
+}
+
+/// Returns true if `s` contains `code` as a standalone number, i.e. not
+/// embedded in a longer digit run. Prevents ports, slots, or request ids from
+/// being mistaken for HTTP status codes (e.g. "127.0.0.1:40329" contains "403").
+fn contains_status_code(s: &str, code: &str) -> bool {
+    let bytes = s.as_bytes();
+    let mut start = 0;
+    while let Some(pos) = s[start..].find(code) {
+        let i = start + pos;
+        let end = i + code.len();
+        let before_ok = i == 0 || !bytes[i - 1].is_ascii_digit();
+        let after_ok = end >= bytes.len() || !bytes[end].is_ascii_digit();
+        if before_ok && after_ok {
+            return true;
+        }
+        start = i + 1;
+    }
+    false
+}
+
+/// Helper to identify whether an RPC error should be retried.
+/// Protects against endlessly retrying terminal client errors (4xx).
+pub fn is_retryable(e: &anyhow::Error) -> bool {
+    let s = e.to_string();
+    // 408 Request Timeout and 429 Too Many Requests are retryable.
+    if contains_status_code(&s, "408") || contains_status_code(&s, "429") {
+        return true;
+    }
+    // Other 4xx errors are generally client errors and should not be retried.
+    if contains_status_code(&s, "400")
+        || contains_status_code(&s, "401")
+        || contains_status_code(&s, "403")
+        || contains_status_code(&s, "404")
+        || contains_status_code(&s, "405")
+    {
+        return false;
+    }
+    true
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn client_errors_are_not_retryable() {
+        for code in ["400", "401", "403", "404", "405"] {
+            let e = anyhow::anyhow!(
+                "HTTP status client error ({code} Some Reason) for url (http://127.0.0.1:1234/)"
+            );
+            assert!(!is_retryable(&e), "{code} should not be retryable");
+        }
+    }
+
+    #[test]
+    fn shared_backoff_rate_limit_cooldown_grows_and_caps() {
+        let sb =
+            SharedBackoff::with_cooldowns(Duration::from_millis(100), Duration::from_millis(400));
+        assert_eq!(sb.report_rate_limited(), Duration::from_millis(100));
+        assert_eq!(sb.report_rate_limited(), Duration::from_millis(200));
+        assert_eq!(sb.report_rate_limited(), Duration::from_millis(400));
+        assert_eq!(sb.report_rate_limited(), Duration::from_millis(400));
+    }
+
+    #[test]
+    fn shared_backoff_shorter_cooldown_does_not_shorten_window() {
+        let sb = SharedBackoff::with_cooldowns(Duration::from_millis(100), Duration::from_secs(60));
+        sb.report_rate_limited();
+        let long = sb.report_rate_limited();
+        sb.report_success();
+        let short = sb.report_rate_limited();
+        assert!(short < long);
+        assert!(sb.remaining() > short);
+    }
+
+    #[test]
+    fn shared_backoff_success_resets_penalty() {
+        let sb = SharedBackoff::with_cooldowns(Duration::from_millis(100), Duration::from_secs(60));
+        sb.report_rate_limited();
+        sb.report_rate_limited();
+        sb.report_success();
+        assert_eq!(sb.report_rate_limited(), Duration::from_millis(100));
+    }
+
+    #[test]
+    fn detects_rate_limit_errors() {
+        assert!(is_rate_limited(&anyhow::anyhow!(
+            "HTTP status client error (429 Too Many Requests) for url (http://x/)"
+        )));
+        // Port 42900 contains "429"; the 500 must not be mistaken for rate limiting.
+        assert!(!is_rate_limited(&anyhow::anyhow!(
+            "HTTP status server error (500) for url (http://127.0.0.1:42900/)"
+        )));
+        assert!(!is_rate_limited(&anyhow::anyhow!(
+            "HTTP status client error (403 Forbidden)"
+        )));
+    }
+
+    #[tokio::test]
+    async fn wait_blocks_until_cooldown_elapses() {
+        let sb = SharedBackoff::with_cooldowns(Duration::from_millis(80), Duration::from_secs(60));
+        let t = std::time::Instant::now();
+        sb.wait().await;
+        assert!(t.elapsed() < Duration::from_millis(50));
+
+        sb.report_rate_limited();
+        let t = std::time::Instant::now();
+        sb.wait().await;
+        assert!(t.elapsed() >= Duration::from_millis(80));
+    }
+
+    #[test]
+    fn retryable_errors() {
+        for msg in [
+            // Regression: port 40329 contains "403"; the 500 must still be retried.
+            "HTTP status server error (500 Internal Server Error) for url (http://127.0.0.1:40329/)",
+            // Digits embedded in longer numbers (ports, slots) are not status codes.
+            "server error (500) for url (http://host:14290/)",
+            "error at slot 14005",
+            // 408/429 are explicitly retryable.
+            "HTTP status client error (408 Request Timeout)",
+            "HTTP status client error (429 Too Many Requests)",
+        ] {
+            assert!(is_retryable(&anyhow::anyhow!(msg)), "{msg:?} should be retryable");
+        }
+    }
+}
